[PATCH] Show patch content in description edit buffer
Add an option (-s/--showpatch) to show the patch content when editing the
patch description, by appending it to the edit buffer, for more accurate
patch comment editing and to check its correctness before committing it.

It's how one works with quilt, and it's very important when the patch is
non-trivial to re-read it.

It's done for import, new and refresh, I left fold out.

Additionally, set the buffer filetype to "diff" for correct highlighting in
Vi.

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade@yahoo.it>