[PATCH] Really fix import --edit invoking editor twice
A typo prevented the previous patch for this (commit
9d15ccd85e36ef5cd554aaf98f1c37ef0433697f) from working (edit is never used,
you should have set options.edit). Ah, nice Python! No var declarations and
these are the effects (yes, I reproduced this).

However, if the description is left empty, refresh_patch will again invoke
the editor. Also, with --showpatch, the first invocation shows the topmost
patch, not the one to be imported.

So better stop new_patch from invoking the editor, and call it only on
refresh.

Btw, clean the old unused "edit" parameter of Series.new_patch.

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade@yahoo.it>