Using delete_patch while deleting a branch is broken
"stg branch --force --delete" takes forever, and often leaves the
repository in a broken state.  The reason for this is the use of
stack.delete_patch() to delete the patches before removing the series.

The problem is delete_patch() will try to pop the patches, which changes
the head link.  After the branch deletion is complete, HEAD will point to
.git/refs/heads/current-branch, but the contents of current-branch will be
the old base of the deleted branch.

Signed-off-by: Chuck Lever <cel@netapp.com>