remove formatting of git log entry on an imported patch
The latest stgit changeset appears to reformat patch descriptions on
import (stripping whitespace and adding newlines). I personally much
prefer to be able to control the git log entry myself rather than have a
program dictate the format to me.

The patch also adds support for Date: and From: lines both before and
after the subject (that is a less important change, and can be split out
or thrown out if you prefer).

Signed-off-by: Trond Myklebust <trond.myklebust@fys.uio.no>