Remove the --force option for new
The default behaviour is now the same as the previous 'new --force' since
most of the people (and myself) first change the tree and run 'new'
afterwards to create the patch. Note that the newly created patch is empty
even if there are local changes. A 'refresh' command is needed to add the
local changes to the patch.

Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>