Add the --update option to pick
This option updates the current patch with changes from the picked commit
or patch. This option is useful when work is done on a temporary patch but
the changes need to be folded back into existing patches. Note that
three-way merging is not supported with this option (I'm not sure it is
even needed).

Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>