Don't always use git-merge-recursive because of speed
There are situations like picking patches, folding a patch onto a base
or synchronising patches where the distance between base and heads is
in general small. In this case, git-read-tree --aggressive is much
faster. We currently use the recursive merge when picking or pushing a
patch (the base change might involve some renames).

Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>