Track the files conflict history
This commit refreshes a pushed patch even if there was a conflict (in
which case only the correctly merged files are merged). The refresh
after fixing the conflict would log the conflict files in the patch
history.

The patch also updates the index in the case of the interactive merge.

Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>