Use branch.*.stgit.parentbranch for authoritative parent information.
Also set the API to use remote=None as "parent branch is local", and
have set_parent() translate that to branch.*.remote=".".

Also enforces that we never set a parent-branch info without a
parent-remote (so they are now both set or both untouched), and try to
give useful info when we have to use the old "origin" defaults.

Signed-off-by: Yann Dirson <ydirson@altern.org>