Fix the importing of multipart emails
The old version was assuming that there are only two parts - the
description and the diff. However, it was failing for PGP-signed or
other multipart encoded emails. This patch generates a full text from
all the text/plain parts and parses it afterwards for description and
diff. This solves the combination of inline or attached patches
together with PGP-signed messages.

Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>