Use --force to overwrite python files
Installing StGIT with "make install" is not working properly if another
version is installed.  Files under ~/lib/python2.5/site-packages/stgit
are not updated.

This causes unexpected problems if switching between different branches
of StGIT.

Using "--force" ensures that the files in the tree are installed even if
the already installed files are newer.

Signed-off-by: Pavel Roskin <proski@gnu.org>
Signed-off-by: Karl Hasselström <kha@treskal.com>