Let "stg status" ignore empty directories
This was a really simple fix: just pass the right flag to
git-ls-files. Since the output has a trailing \0, split() gives us an
empty string at the end of the list that we have to throw away.
(Curiously, there was an empty string at the end before this change
too, but it disappeared somehow.)

This fixes bug 9891.

Signed-off-by: Karl Hasselström <kha@treskal.com>