Fix the rebasing with an external command
Since git.py caches the HEAD value, rebasing with and external command
didn't clear the cache and the subsequent push simple fast-forwards
the tree. The patch creates a new git.rebase() function which takes
care of custom rebase commands. It also removes "must_rebase" in
pull.py as it doesn't seem to be used.

Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>