Do not raise an exception if no FETCH_HEAD
The fetchcmd option in .git/config can be set to a command that
doesn't generate a FETCH_HEAD file (like 'git svn fetch'). With this
patch, StGIT will only warn if no FETCH_HEAD is found as it is quite
likely that the rebasecmd doesn't need one ('git svn rebase'). The
patch also modifies the 'fetch-rebase' case in pull.py so that it
re-raises the exception initially raised by git.fetch_head() as this
function has the correct information about the type of exception.

Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>