Write removed fields for backwards compatibility
Start writing the "top", "bottom", and "bottom.old" fields again. The
last two patches mean we don't need them anymore, but old versions of
StGit still do.

At some later time, we'll want to change the stack on-disk format so
that these can be removed. But we'll probably want to do that as part
of a larger change.

Signed-off-by: Karl Hasselström <kha@treskal.com>