Don't assume there is always a stage 2 in git.get_conflicts
For example, the t1202-push-undo.sh test generates a conflict where a
file was added in both current and patch but different content and
missing in ancestor, therefore no stage 2.

There could also be a case where stage 3 is missing if a file is
removed by the patch being pushed.

Signed-off-by: Catalin Marinas <catalin.marinas@gmail.com>