Add an --index option to "stg refresh"
Add an --index option to "stg refresh" which takes the contents of the
index as the new commit.

This allows to stage only certain changes to a file by only adding the
desired parts to the index with git-gui, ugit, git add -i or another
tool that manipulates the index and then run stg refresh --index it.
Also allows removing a file from a patch by running git reset HEAD^ --
file_to_remove followed by a stg refresh --index.

Signed-off-by: Peter Oberndorfer <kumbayo84@arcor.de>
Signed-off-by: Karl Hasselström <kha@treskal.com>